# dashboard/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: MONOLITH OS API
  description: |
    RESTful API for MONOLITH OS - An AI-powered enterprise operating system that orchestrates
    autonomous agents for business operations, decision-making, and workflow management.

    ## Features
    - **Task Management**: Create, update, and track tasks across the organization
    - **Decision Logging**: Record and audit business decisions with full traceability
    - **Workflow Orchestration**: Manage multi-step business processes
    - **Agent Management**: Monitor and control AI agents across different roles

    ## Authentication
    All endpoints require Bearer token authentication. Obtain a token via `/auth/login`.

    ## Rate Limiting
    - Standard endpoints: 100 requests/minute
    - Bulk operations: 10 requests/minute

    ## Versioning
    The API uses URL-based versioning. Current version: v1
  version: 1.0.0
  contact:
    name: MONOLITH OS Support
    email: support@monolith-os.local
  license:
    name: Proprietary
    url: https://monolith-os.local/license
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.monolith-os.local/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /v1/decisions:
    get:
      tags:
        - Decisions
      summary: List decisions
      description: Retrieve a paginated list of decisions with optional filtering
      parameters:
        - name: role
          in: query
          description: Filter by decision maker role
          schema:
            type: string
        - name: category
          in: query
          description: Filter by category
          schema:
            type: string
            enum: [budget, strategy, compliance, security, operations, hr, marketing, legal, technology]
        - name: impact
          in: query
          description: Filter by impact level
          schema:
            type: string
            enum: [low, medium, high, critical]
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/LimitParam'
      responses:
        '200':
          description: List of decisions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionListResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

    post:
      tags:
        - Decisions
      summary: Record a decision
      description: Record a new business decision
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionCreate'
      responses:
        '201':
          description: Decision recorded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
  parameters:
...
    PageParam:
      name: page
      in: query
      description: Page number
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      description: Items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
  schemas:
...
    Decision:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        decided_at:
          type: string
          format: date-time
        decided_by_role:
          type: string
        decided_by_name:
          type: string
        category:
          type: string
          enum: [budget, strategy, compliance, security, operations, hr, marketing, legal, technology]
        impact:
          type: string
          enum: [low, medium, high, critical]
        financial_impact:
          type: number
          nullable: true
    DecisionCreate:
      type: object
      required:
        - decision
      properties:
        decision:
          type: string
          description: Decision description
        role:
          type: string
          description: Decision maker role
        task_id:
          type: string
          description: Associated task ID
        rationale:
          type: string
          description: Decision rationale
        financial_impact:
          type: number
          description: Financial impact amount
        category:
          type: string
          enum: [budget, strategy, compliance, security, operations, hr, marketing, legal, technology]
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
        hasNext:
          type: boolean
        hasPrev:
          type: boolean
...
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            field:
              type: string
...
    DecisionListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Decision'
        pagination:
          $ref: '#/components/schemas/Pagination'
    DecisionResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          $ref: '#/components/schemas/Decision'
        message:
          type: string
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid input
              field: content
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred